iris/requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
blessings==1.6
bootstrap-admin==0.3.6
bpython==0.14.2
certifi==2016.9.26
configparser==3.5.0
curtsies==0.1.23
decorator==4.0.10
diff-match-patch==20121119
Django==1.8.18
django-audit-log==0.7.0
django-bootstrap-breadcrumbs==0.8.2
django-bootstrap3==7.0.1
django-clear-cache==0.3
django-extensions==1.7.2
django-import-export==0.4.5
django-localflavor==1.2
django-model-utils==2.4
django-report-builder==3.3.0
django-report-utils==0.3.15
django-smart-selects==1.5.2
django-suit==0.2.16
djangorestframework==3.3.3
docutils==0.12
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
greenlet==0.4.12
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
jdcal==1.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.2.0
notebook==4.2.3
olefile==0.44
openpyxl==2.2.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.0.2
python-dateutil==2.5.3
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.9.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sorl-thumbnail==12.4a1
tablib==0.11.2
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
